Add support for Service VM to manage NVSwitches
A dedicated Service VM can be used to manage the NVSwitches.
The NVSwitches are passed through to the Service VM.
The Fabric Manager runs in the Service VM and is configured
to listen on the Service VM's network interface IP.
The libvirt hook script for QEMU on the host uses the
Fabric Manager Partition Manager Client to communicate to the
Fabric Manager.  It needs to be updated to used the IP the
Fabric Manager is configured to listen on in the Service VM.